requirements_.txt
streamlit==1.38.0
langchain-google-genai==1.0.10
google-generativeai==0.7.2
langchain==0.2.14
langchain-core==0.2.32
python-dotenv
